Improve type safety
Type safety is completely optional (and I cannot see how to do it
better with current TypeScript map-on-types capabilities).  So we
keep the relatively type-unsafe one-arg `by` but allow the caller
to specify a particular return type.  That type is checked at run
time for compatibility -- for efficiency simply call `by` a fixed
number of times before starting any loops.